requirements.txt
fastapi
google-genai